🌱 [scanner] refactor: split SecretDrillDown and ComplianceDrillDown (#21791)
- SecretDrillDown.tsx: 457→112 lines (hooks→useSecretDrillDown.ts, UI→SecretDrillDown.parts.tsx)
- ComplianceDrillDown.tsx: 515→240 lines (UI chunks→ComplianceDrillDown.parts.tsx)
- No behaviour change; all exports preserved
