+++ web/src/components/drilldown/views/SecretDrillDown.tsx
@@ -1,20 +1,3 @@
-import { useState, useEffect, useRef } from 'react'
-import { useLocalAgent } from '../../../hooks/useLocalAgent'
-import { useDrillDownWebSocket } from '../../../hooks/useDrillDownWebSocket'
-import { useDrillDownActions, useDrillDown } from '../../../hooks/useDrillDown'
-import { ClusterBadge } from '../../ui/ClusterBadge'
-import { FileText, Code, Info, Tag, ChevronDown, ChevronUp, Loader2, Copy, Check, ChevronLeft, Layers, Server, Eye, EyeOff, Lock } from 'lucide-react'
-import { cn } from '../../../lib/cn'
-import { UI_FEEDBACK_TIMEOUT_MS } from '../../../lib/constants/network'
-import { useTranslation } from 'react-i18next'
-import { copyToClipboard } from '../../../lib/clipboard'
-
-interface Props {
-  data: Record<string, unknown>
-}
-
-type TabType = 'overview' | 'data' | 'describe' | 'yaml'
-
 // #6231: the regex-based maskSecretYaml that used to live here had two
 // real bugs (block-scalar handling, false bundle-bloat claim about
 // js-yaml). Replaced by a shared js-yaml-based helper in lib/yamlMask.
@@ -25,431 +8,103 @@ import { maskKubernetesYamlData } from '../../../lib/yamlMask'
 /** @deprecated use `maskKubernetesYamlData` from `lib/yamlMask` */
 export const maskSecretYaml = maskKubernetesYamlData
 
-/** Property names that must never be used as object keys (prototype pollution prevention). */
-const UNSAFE_PROP_NAMES = new Set(['__proto__', 'constructor', 'prototype'])
+import { useSecretDrillDown } from './useSecretDrillDown'
+import {
+  SecretHeaderBreadcrumbs,
+  SecretTabBar,
+  SecretOverviewTab,
+  SecretDataTab,
+  SecretDescribeTab,
+  SecretYamlTab,
+} from './SecretDrillDown.parts'
+
+interface Props {
+  data: Record<string, unknown>
+}
 
 export function SecretDrillDown({ data }: Props) {
-  const { t } = useTranslation()
-  const cluster = data.cluster as string
-  const namespace = data.namespace as string
-  const secretName = data.secret as string
-  const { isConnected: agentConnected } = useLocalAgent()
-  const { drillToNamespace, drillToCluster } = useDrillDownActions()
-  const { state, pop } = useDrillDown()
-  const { runKubectl } = useDrillDownWebSocket(cluster)
-
-  const [activeTab, setActiveTab] = useState<TabType>('overview')
-  const [secretData, setSecretData] = useState<Record<string, string> | null>(null)
-  const [secretType, setSecretType] = useState<string | null>(null)
-  const [describeOutput, setDescribeOutput] = useState<string | null>(null)
-  const [describeLoading, setDescribeLoading] = useState(false)
-  const [yamlOutput, setYamlOutput] = useState<string | null>(null)
-  const [yamlLoading, setYamlLoading] = useState(false)
-  const [dataLoading, setDataLoading] = useState(false)
-  const [dataError, setDataError] = useState<string | null>(null)
-  const [labels, setLabels] = useState<Record<string, string> | null>(null)
-  const [copiedField, setCopiedField] = useState<string | null>(null)
-  const copiedFieldTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-  const [showAllData, setShowAllData] = useState(false)
-  const [revealedKeys, setRevealedKeys] = useState<Set<string>>(new Set())
-  // YAML tab reveal state — defaults to MASKED so the `data:` block doesn't
-  // leak base64-encoded secrets the moment a user clicks the YAML tab
-  // (#6209). Mirrors the per-key reveal pattern on the Data tab — explicit
-  // user action required to see secrets.
-  const [yamlRevealed, setYamlRevealed] = useState(false)
-
-
-  // Fetch Secret data
-  const fetchData = async () => {
-    if (!agentConnected) return
-    setDataLoading(true)
-    setDataError(null)
-    try {
-      const output = await runKubectl(['get', 'secret', secretName, '-n', namespace, '-o', 'json'])
-      if (output) {
-        const secret = JSON.parse(output)
-        // Decode base64 data (use null-prototype object to prevent prototype pollution)
-        const decodedData: Record<string, string> = Object.create(null) as Record<string, string>
-        if (secret.data) {
-          for (const [key, value] of Object.entries(secret.data)) {
-            if (UNSAFE_PROP_NAMES.has(key)) continue
-            try {
-              decodedData[key] = atob(value as string)
-            } catch {
-              decodedData[key] = value as string
-            }
-          }
-        }
-        setSecretData(decodedData)
-        setSecretType(secret.type || 'Opaque')
-        setLabels(secret.metadata?.labels || {})
-      }
-    } catch (err) {
-      setDataError(err instanceof Error ? err.message : t('common.fetchFailed', 'Failed to load Secret data'))
-    } finally {
-      setDataLoading(false)
-    }
-  }
-
-  const fetchDescribe = async () => {
-    if (!agentConnected || describeOutput) return
-    setDescribeLoading(true)
-    const output = await runKubectl(['describe', 'secret', secretName, '-n', namespace])
-    setDescribeOutput(output)
-    setDescribeLoading(false)
-  }
-
-  const fetchYaml = async () => {
-    if (!agentConnected || yamlOutput) return
-    setYamlLoading(true)
-    const output = await runKubectl(['get', 'secret', secretName, '-n', namespace, '-o', 'yaml'])
-    setYamlOutput(output)
-    setYamlLoading(false)
-  }
-
-  // Track if we've already loaded data to prevent refetching
-  const hasLoadedRef = useRef(false)
-
-  // Pre-fetch tab data when agent connects
-  // Batched to limit concurrent WebSocket connections (max 2 at a time)
-  useEffect(() => {
-    if (!agentConnected || hasLoadedRef.current) return
-    hasLoadedRef.current = true
-
-    const loadData = async () => {
-      // Batch 1: Overview data (2 concurrent)
-      await Promise.all([
-        fetchData(),
-        fetchDescribe(),
-      ])
-
-      // Batch 2: YAML (lower priority)
-      await fetchYaml()
-    }
-
-    loadData()
-  }, [agentConnected, fetchData, fetchDescribe, fetchYaml])
-
-  useEffect(() => {
-    return () => {
-      if (copiedFieldTimeoutRef.current) {
-        clearTimeout(copiedFieldTimeoutRef.current)
-      }
-    }
-  }, [])
-
-  const handleCopy = (field: string, value: string) => {
-    copyToClipboard(value)
-    setCopiedField(field)
-    if (copiedFieldTimeoutRef.current) {
-      clearTimeout(copiedFieldTimeoutRef.current)
-    }
-    copiedFieldTimeoutRef.current = setTimeout(() => {
-      setCopiedField(null)
-      copiedFieldTimeoutRef.current = null
-    }, UI_FEEDBACK_TIMEOUT_MS)
-  }
-
-  const toggleReveal = (key: string) => {
-    setRevealedKeys(prev => {
-      const next = new Set(prev)
-      if (next.has(key)) {
-        next.delete(key)
-      } else {
-        next.add(key)
-      }
-      return next
-    })
-  }
-
-  const TABS: { id: TabType; label: string; icon: typeof Info }[] = [
-    { id: 'overview', label: t('drilldown.tabs.overview', 'Overview'), icon: Info },
-    { id: 'data', label: t('drilldown.tabs.data', 'Data'), icon: Lock },
-    { id: 'describe', label: t('drilldown.tabs.describe', 'Describe'), icon: FileText },
-    { id: 'yaml', label: t('drilldown.tabs.yaml', 'YAML'), icon: Code },
-  ]
-
-  const dataEntries = Object.entries(secretData || {})
-  const displayedData = showAllData ? dataEntries : dataEntries.slice(0, 5)
+  const {
+    cluster,
+    namespace,
+    secretName,
+    drillToNamespace,
+    drillToCluster,
+    stackLength,
+    pop,
+    activeTab,
+    setActiveTab,
+    tabs,
+    secretType,
+    dataLoading,
+    dataError,
+    describeOutput,
+    describeLoading,
+    yamlOutput,
+    yamlLoading,
+    yamlRevealed,
+    toggleYamlRevealed,
+    copiedField,
+    handleCopy,
+    showAllData,
+    setShowAllData,
+    revealedKeys,
+    toggleReveal,
+    dataEntries,
+    displayedData,
+    labels,
+  } = useSecretDrillDown(data)
 
   return (
     <div className="flex flex-col h-full -m-6">
-      {/* Header */}
-      <div className="px-6 pt-6 pb-4">
-        <div className="flex items-center gap-6 text-sm">
-          {state.stack.length > 1 && (
-            <button
-              type="button"
-              onClick={pop}
-              className="flex items-center gap-2 hover:bg-secondary/50 border border-transparent hover:border-border px-3 py-1.5 rounded-lg transition-all text-muted-foreground hover:text-foreground"
-              aria-label={t('drilldown.goBack')}
-              title={t('drilldown.goBack')}
-            >
-              <ChevronLeft className="w-4 h-4" />
-              <span>{t('common.back')}</span>
-            </button>
-          )}
-          <button
-            type="button"
-            onClick={() => drillToNamespace(cluster, namespace)}
-            className="flex items-center gap-2 hover:bg-purple-500/10 border border-transparent hover:border-purple-500/30 px-3 py-1.5 rounded-lg transition-all group cursor-pointer"
-          >
-            <Layers className="w-4 h-4 text-purple-400" />
-            <span className="text-muted-foreground">{t('drilldown.fields.namespace')}</span>
-            <span className="font-mono text-purple-400 group-hover:text-purple-300 transition-colors">{namespace}</span>
-            <svg className="w-3 h-3 text-purple-400/70 group-hover:text-purple-400 transition-colors" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-            </svg>
-          </button>
-          <button
-            type="button"
-            onClick={() => drillToCluster(cluster)}
-            className="flex items-center gap-2 hover:bg-blue-500/10 border border-transparent hover:border-blue-500/30 px-3 py-1.5 rounded-lg transition-all group cursor-pointer"
-          >
-            <Server className="w-4 h-4 text-blue-400" />
-            <span className="text-muted-foreground">{t('drilldown.fields.cluster')}</span>
-            <ClusterBadge cluster={cluster.split('/').pop() || cluster} size="sm" />
-            <svg className="w-3 h-3 text-blue-400/70 group-hover:text-blue-400 transition-colors" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-            </svg>
-          </button>
-        </div>
-      </div>
-
-      {/* Tabs */}
-      <div className="border-b border-border px-6">
-        <div className="flex gap-1">
-          {TABS.map((tab) => {
-            const Icon = tab.icon
-            return (
-              <button
-                key={tab.id}
-                onClick={() => setActiveTab(tab.id)}
-                className={cn(
-                  'px-4 py-2 text-sm font-medium flex items-center gap-2 border-b-2 transition-colors',
-                  activeTab === tab.id
-                    ? 'text-primary border-primary'
-                    : 'text-muted-foreground border-transparent hover:text-foreground hover:border-border'
-                )}
-              >
-                <Icon className="w-4 h-4" />
-                {tab.label}
-              </button>
-            )
-          })}
-        </div>
-      </div>
-
-      {/* Tab Content */}
+      <SecretHeaderBreadcrumbs
+        cluster={cluster}
+        namespace={namespace}
+        canGoBack={stackLength > 1}
+        onBack={pop}
+        onNamespaceClick={() => drillToNamespace(cluster, namespace)}
+        onClusterClick={() => drillToCluster(cluster)}
+      />
+      <SecretTabBar tabs={tabs} activeTab={activeTab} onTabChange={setActiveTab} />
       <div className="flex-1 overflow-y-auto p-6 space-y-6">
         {activeTab === 'overview' && (
-          <div className="space-y-6">
-            {dataLoading && (
-              <div className="flex items-center justify-center py-6">
-                <Loader2 className="w-5 h-5 animate-spin text-primary" />
-                <span className="ml-2 text-sm text-muted-foreground">{t('common.loading', 'Loading...')}</span>
-              </div>
-            )}
-            {dataError && !dataLoading && (
-              <div className="p-3 rounded-lg bg-red-500/10 border border-red-500/20">
-                <p className="text-sm text-red-400">{dataError}</p>
-              </div>
-            )}
-            {/* Basic Info */}
-            <div className="p-4 rounded-lg bg-red-500/10 border border-red-500/20">
-              <div className="flex items-center gap-3">
-                <Lock className="w-8 h-8 text-red-400" />
-                <div>
-                  <div className="text-lg font-semibold text-foreground">{secretName}</div>
-                  <div className="text-sm text-muted-foreground">
-                    Type: {secretType} • {dataEntries.length} key{dataEntries.length !== 1 ? 's' : ''}
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            {/* Labels */}
-            {labels && Object.keys(labels).length > 0 && (
-              <div>
-                <h3 className="text-sm font-semibold text-foreground mb-2 flex items-center gap-2">
-                  <Tag className="w-4 h-4 text-blue-400" />
-                  Labels
-                </h3>
-                <div className="flex flex-wrap gap-2">
-                  {Object.entries(labels).slice(0, 5).map(([key, value]) => (
-                    <span key={key} className="text-xs px-2 py-1 rounded bg-blue-500/10 text-blue-400 font-mono">
-                      {key}={value}
-                    </span>
-                  ))}
-                  {Object.keys(labels).length > 5 && (
-                    <span className="text-xs text-muted-foreground">+{Object.keys(labels).length - 5} more</span>
-                  )}
-                </div>
-              </div>
-            )}
-
-            {/* Data Keys */}
-            <div>
-              <h3 className="text-sm font-semibold text-foreground mb-2 flex items-center gap-2">
-                <Lock className="w-4 h-4 text-red-400" />
-                Secret Keys
-              </h3>
-              {dataEntries.length > 0 ? (
-                <div className="flex flex-wrap gap-2">
-                  {dataEntries.map(([key]) => (
-                    <span key={key} className="text-xs px-2 py-1 rounded bg-red-500/10 text-red-400 font-mono">
-                      {key}
-                    </span>
-                  ))}
-                </div>
-              ) : (
-                <p className="text-sm text-muted-foreground">No data in this Secret</p>
-              )}
-            </div>
-          </div>
+          <SecretOverviewTab
+            secretName={secretName}
+            secretType={secretType}
+            dataLoading={dataLoading}
+            dataError={dataError}
+            dataEntries={dataEntries}
+            labels={labels}
+          />
         )}
-
         {activeTab === 'data' && (
-          <div className="space-y-4">
-            <div className="p-3 rounded-lg bg-yellow-500/10 border border-yellow-500/20 text-sm text-yellow-400 flex items-center gap-2">
-              <Lock className="w-4 h-4" />
-              Secret values are hidden by default. Click the eye icon to reveal.
-            </div>
-            {dataEntries.length > 0 ? (
-              <>
-                {displayedData.map(([key, value]) => (
-                  <div key={key} className="rounded-lg bg-card/50 border border-border overflow-hidden">
-                    <div className="flex items-center justify-between px-3 py-2 bg-red-500/10 border-b border-border">
-                      <span className="font-mono text-sm text-red-400">{key}</span>
-                      <div className="flex items-center gap-1">
-                        <button
-                          onClick={() => toggleReveal(key)}
-                          className="p-1 rounded hover:bg-secondary/50 text-muted-foreground hover:text-foreground"
-                        >
-                          {revealedKeys.has(key) ? (
-                            <EyeOff className="w-4 h-4" />
-                          ) : (
-                            <Eye className="w-4 h-4" />
-                          )}
-                        </button>
-                        <button
-                          onClick={() => handleCopy(`data-${key}`, value)}
-                          className="p-1 rounded hover:bg-secondary/50 text-muted-foreground hover:text-foreground"
-                        >
-                          {copiedField === `data-${key}` ? (
-                            <Check className="w-4 h-4 text-green-400" />
-                          ) : (
-                            <Copy className="w-4 h-4" />
-                          )}
-                        </button>
-                      </div>
-                    </div>
-                    <pre className="p-3 text-xs font-mono text-foreground whitespace-pre-wrap max-h-48 overflow-auto">
-                      {revealedKeys.has(key) ? value : '••••••••••••••••'}
-                    </pre>
-                  </div>
-                ))}
-                {dataEntries.length > 5 && (
-                  <button
-                    onClick={() => setShowAllData(!showAllData)}
-                    className="text-xs text-primary hover:text-primary/80 flex items-center gap-1"
-                  >
-                    {showAllData ? (
-                      <>Show less <ChevronUp className="w-3 h-3" /></>
-                    ) : (
-                      <>Show all {dataEntries.length} keys <ChevronDown className="w-3 h-3" /></>
-                    )}
-                  </button>
-                )}
-              </>
-            ) : (
-              <div className="p-4 rounded-lg bg-card/50 border border-border text-center text-muted-foreground">
-                No data in this Secret
-              </div>
-            )}
-          </div>
+          <SecretDataTab
+            dataEntries={dataEntries}
+            displayedData={displayedData}
+            revealedKeys={revealedKeys}
+            copiedField={copiedField}
+            showAllData={showAllData}
+            onToggleReveal={toggleReveal}
+            onCopy={handleCopy}
+            onToggleShowAll={() => setShowAllData(!showAllData)}
+          />
         )}
-
         {activeTab === 'describe' && (
-          <div>
-            {describeLoading ? (
-              <div className="flex items-center justify-center py-12">
-                <Loader2 className="w-6 h-6 animate-spin text-primary" />
-                <span className="ml-2 text-muted-foreground">{t('drilldown.status.runningDescribe')}</span>
-              </div>
-            ) : describeOutput ? (
-              <div className="relative">
-                <button
-                  onClick={() => handleCopy('describe', describeOutput)}
-                  className="absolute top-2 right-2 px-2 py-1 rounded bg-secondary/50 text-xs text-muted-foreground hover:text-foreground flex items-center gap-1"
-                >
-                  {copiedField === 'describe' ? <><Check className="w-3 h-3 text-green-400" /> Copied</> : <><Copy className="w-3 h-3" /> Copy</>}
-                </button>
-                <pre className="p-4 rounded-lg bg-black/50 border border-border overflow-auto max-h-[60vh] text-xs text-foreground font-mono whitespace-pre-wrap">
-                  {describeOutput}
-                </pre>
-              </div>
-            ) : (
-              <div className="p-4 rounded-lg bg-yellow-500/10 border border-yellow-500/20 text-center">
-                <p className="text-yellow-400">{t('drilldown.empty.localAgentNotConnected')}</p>
-              </div>
-            )}
-          </div>
+          <SecretDescribeTab
+            describeLoading={describeLoading}
+            describeOutput={describeOutput}
+            copiedField={copiedField}
+            onCopy={handleCopy}
+          />
         )}
-
         {activeTab === 'yaml' && (
-          <div className="space-y-3">
-            {/* #6209: same warning the Data tab uses, so the YAML tab is no
-                longer the easy bypass for the per-key reveal model. */}
-            <div className="p-3 rounded-lg bg-yellow-500/10 border border-yellow-500/20 text-sm text-yellow-400 flex items-center gap-2">
-              <Lock className="w-4 h-4" />
-              {yamlRevealed
-                ? 'Secret values are visible. Click the eye icon to mask.'
-                : 'Secret values are hidden by default. Click the eye icon to reveal.'}
-            </div>
-            {yamlLoading ? (
-              <div className="flex items-center justify-center py-12">
-                <Loader2 className="w-6 h-6 animate-spin text-primary" />
-                <span className="ml-2 text-muted-foreground">{t('drilldown.status.fetchingYaml')}</span>
-              </div>
-            ) : yamlOutput ? (
-              <div className="relative">
-                {(() => {
-                  // Compute once per render so the Copy button puts the
-                  // SAME bytes onto the clipboard that the user is looking
-                  // at — masked when masked, raw when revealed (#6209).
-                  const displayedYaml = yamlRevealed ? yamlOutput : maskSecretYaml(yamlOutput)
-                  return (
-                    <>
-                      <div className="absolute top-2 right-2 flex items-center gap-1">
-                        <button
-                          onClick={() => setYamlRevealed(v => !v)}
-                          className="p-1 rounded bg-secondary/50 text-muted-foreground hover:text-foreground"
-                          aria-label={yamlRevealed ? 'Mask secret values' : 'Reveal secret values'}
-                          title={yamlRevealed ? 'Mask secret values' : 'Reveal secret values'}
-                        >
-                          {yamlRevealed ? <EyeOff className="w-4 h-4" /> : <Eye className="w-4 h-4" />}
-                        </button>
-                        <button
-                          onClick={() => handleCopy('yaml', displayedYaml)}
-                          className="px-2 py-1 rounded bg-secondary/50 text-xs text-muted-foreground hover:text-foreground flex items-center gap-1"
-                        >
-                          {copiedField === 'yaml' ? <><Check className="w-3 h-3 text-green-400" /> Copied</> : <><Copy className="w-3 h-3" /> Copy</>}
-                        </button>
-                      </div>
-                      <pre className="p-4 rounded-lg bg-black/50 border border-border overflow-auto max-h-[60vh] text-xs text-foreground font-mono whitespace-pre-wrap">
-                        {displayedYaml}
-                      </pre>
-                    </>
-                  )
-                })()}
-              </div>
-            ) : (
-              <div className="p-4 rounded-lg bg-yellow-500/10 border border-yellow-500/20 text-center">
-                <p className="text-yellow-400">{t('drilldown.empty.localAgentNotConnected')}</p>
-              </div>
-            )}
-          </div>
+          <SecretYamlTab
+            yamlLoading={yamlLoading}
+            yamlOutput={yamlOutput}
+            yamlRevealed={yamlRevealed}
+            copiedField={copiedField}
+            onToggleReveal={toggleYamlRevealed}
+            onCopy={handleCopy}
+          />
         )}
       </div>
     </div>
